+++ js/main.js
@@ -6,11 +6,9 @@ import {
   loadLocalStorageItems,
 } from "./localStorage.js";
 
-// start of elements below
 export let listDiv = document.getElementById("list");
 let userTextInput = document.getElementById("userListEntry");
 let userSubmitButton = document.getElementById("userAddButton");
-// end of elements above
 
 // start of event listeners
 userSubmitButton.addEventListener("click", function () {
